+++ api/package.json
@@ -23,7 +23,6 @@
   "license": "MIT",
   "type": "commonjs",
   "dependencies": {
-    "@aws-sdk/client-sesv2": "^3.968.0",
     "@mybiblelog/shared": "file:../shared",
     "bcrypt": "^6.0.0",
     "compression": "^1.8.1",
@@ -38,11 +37,10 @@
     "http-errors": "^2.0.0",
     "http-status": "^2.1.0",
     "jsonwebtoken": "^9.0.2",
-    "mailgun.js": "^12.1.1",
     "mongoose": "^9.0.0",
     "mongoose-unique-validator": "^5.0.1",
     "morgan": "^1.10.0",
-    "nodemailer": "^7.0.12",
+    "resend": "^6.8.0",
     "supertest": "^7.1.4",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1",
