+++ api/services/email/assets/brand.png
